+++ MaxMind.MinFraud/Exception/AuthenticationException.cs
@@ -32,17 +32,5 @@ namespace MaxMind.MinFraud.Exception
         public AuthenticationException(string message, System.Exception innerException) : base(message, innerException)
         {
         }
-
-        /// <summary>
-        ///     Constructor for deserialization.
-        /// </summary>
-        /// <param name="info"></param>
-        /// <param name="context"></param>
-#if NET8_0_OR_GREATER
-        [Obsolete(DiagnosticId = "SYSLIB0051")]
-#endif
-        protected AuthenticationException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-        }
     }
 }